# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News Aggregator API",
        "description": "API documentation for the News Aggregator application.",
        "contact": {
            "name": "News Aggregator API Support",
            "email": "admin@newsaggregator.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/user/preferences": {
            "get": {
                "tags": [
                    "User Preferences"
                ],
                "summary": "Get user preferences",
                "operationId": "852282c847586808ce870da7a8c58815",
                "responses": {
                    "200": {
                        "description": "User preferences retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sources": {
                                            "description": "List of preferred news sources",
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "CNN",
                                                "BBC",
                                                "The New York Times"
                                            ]
                                        },
                                        "categories": {
                                            "description": "List of preferred news categories",
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "Politics",
                                                "Technology",
                                                "Health"
                                            ]
                                        },
                                        "authors": {
                                            "description": "List of preferred authors to follow",
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "John Smith",
                                                "Jane Doe"
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Preferences not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/preferences/update": {
            "put": {
                "tags": [
                    "User Preferences"
                ],
                "summary": "Update user preferences",
                "operationId": "125ef91cf7928b3a4a5f93589e7c494c",
                "requestBody": {
                    "description": "User preference data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "sources",
                                    "categories",
                                    "authors"
                                ],
                                "properties": {
                                    "sources": {
                                        "description": "List of preferred news sources",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "CNN",
                                            "BBC",
                                            "The New York Times"
                                        ]
                                    },
                                    "categories": {
                                        "description": "List of preferred news categories",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "Politics",
                                            "Technology",
                                            "Health"
                                        ]
                                    },
                                    "authors": {
                                        "description": "List of preferred authors to follow",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "John Smith",
                                            "Jane Doe"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User preferences updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Preferences updated successfully"
                                        },
                                        "preferences": {
                                            "properties": {
                                                "sources": {
                                                    "description": "List of preferred news sources",
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "CNN",
                                                        "BBC",
                                                        "The New York Times"
                                                    ]
                                                },
                                                "categories": {
                                                    "description": "List of preferred news categories",
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "Politics",
                                                        "Technology",
                                                        "Health"
                                                    ]
                                                },
                                                "authors": {
                                                    "description": "List of preferred authors to follow",
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "John Smith",
                                                        "Jane Doe"
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/preferences/reset": {
            "post": {
                "tags": [
                    "User Preferences"
                ],
                "summary": "Reset user preferences to default",
                "operationId": "0138575e7c0a515e4fb7e6a6d10deeb7",
                "responses": {
                    "200": {
                        "description": "User preferences reset successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Preferences reset successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/feed": {
            "get": {
                "tags": [
                    "User Feed"
                ],
                "summary": "Get personalized news feed based on user preferences",
                "operationId": "a26a81808f8fb281fa04c28fe4d51f95",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Personalized articles retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "articles": {
                                            "description": "List of personalized articles based on preferences",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Article"
                                            }
                                        },
                                        "pagination": {
                                            "description": "Pagination information",
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": 45
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "from": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "to": {
                                                    "type": "integer",
                                                    "example": 15
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Article": {
                "title": "Article",
                "required": [
                    "id",
                    "title"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "author": {
                        "type": "string"
                    },
                    "source": {
                        "type": "string"
                    },
                    "category": {
                        "type": "string"
                    },
                    "url": {
                        "type": "string",
                        "format": "uri"
                    },
                    "url_to_image": {
                        "type": "string",
                        "format": "uri"
                    },
                    "published_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token in the format **Bearer {token}**",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}